Split socks5 tests into focused files
